Moved `lib/ApiCaller` to `lib/input/figma/helper`
